@@ -12,16 +12,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.8",
-    "@angular/common": "^14.2.8",
-    "@angular/compiler": "^14.2.8",
-    "@angular/core": "^14.2.8",
-    "@angular/forms": "^14.2.8",
-    "@angular/localize": "^14.2.8",
-    "@angular/platform-browser": "^14.2.8",
-    "@angular/platform-browser-dynamic": "^14.2.8",
-    "@angular/router": "^14.2.8",
-    "@ng-bootstrap/ng-bootstrap": "^13.1.0",
+    "@angular/animations": "^15.2.8",
+    "@angular/common": "^15.2.8",
+    "@angular/compiler": "^15.2.8",
+    "@angular/core": "^15.2.8",
+    "@angular/forms": "^15.2.8",
+    "@angular/localize": "^15.2.8",
+    "@angular/platform-browser": "^15.2.8",
+    "@angular/platform-browser-dynamic": "^15.2.8",
+    "@angular/router": "^15.2.8",
+    "@ng-bootstrap/ng-bootstrap": "^14.1.0",
     "@popperjs/core": "^2.10.2",
     "leaflet": "^1.7.1",
     "bootstrap": "^5.2.2",
@@ -31,9 +31,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.8",
-    "@angular/cli": "^14.2.8",
-    "@angular/compiler-cli": "^14.2.8",
+    "@angular-devkit/build-angular": "^15.2.6",
+    "@angular/cli": "^15.2.6",
+    "@angular/compiler-cli": "^15.2.8",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "@types/leaflet": "^1.7.0",
